Fixed make_confusion()

+++ helper_functions.py
@@ -545,9 +545,9 @@ def oh_to_label(oh_array, dim):
             print(n)
         for i in range(oh_array.shape[1]):
             for j in range(oh_array.shape[2]):
-                holder[n, i, j] = np.asarray(label_dict[tuple(oh_array[n, i, j])])
+                holder[n, i, j] = np.asarray(label_dict[tuple(oh_array[n, i, j])], dtype=np.uint8)
 
-    return holder
+    return holder.astype(np.uint8)
 
 def label_to_oh(label_array, dim):
     arr_list = np.identity(dim, dtype=np.uint8)
@@ -561,9 +561,9 @@ def label_to_oh(label_array, dim):
             print(n)
         for i in range(label_array.shape[1]):
             for j in range(label_array.shape[2]):
-                holder[n, i, j] = np.asarray(label_dict[tuple(label_array[n, i, j])])
+                holder[n, i, j] = np.asarray(label_dict[tuple(label_array[n, i, j])], dtype=np.uint8)
 
-    return holder
+    return holder.astype(np.uint8)
 
 
 def vec_to_oh(array, progress=False, cycle=100):
